% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dom_helpers.R
\name{create_dom_url}
\alias{create_dom_url}
\title{Maak een URL naar de csv domeintabel}
\usage{
create_dom_url(naam, limit = 500, offset = 0)
}
\arguments{
\item{naam}{Naam van de domeintabel}

\item{limit}{Aantal waarden per keer}

\item{offset}{Startpunt}
}
\value{
Een string met de URL
}
\description{
Deze functie genereert de juiste URL van een domeintabel voor de functie \code{dom()}
}
