% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dom_helpers.R
\name{dom_guid}
\alias{dom_guid}
\title{Guid van domeintabel}
\usage{
dom_guid(namen)
}
\arguments{
\item{namen}{Character vector met namen van domeintabellen.}
}
\value{
Vector met guid's
}
\description{
Zoek de guid van domeintabellen op.
}
\examples{
\dontrun{

dom_guid("MonsterType")

}
}
