% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dom_helpers.R
\name{dom_kolommen}
\alias{dom_kolommen}
\title{Kolommen van een domeintabel}
\usage{
dom_kolommen(naam)
}
\arguments{
\item{naam}{Naam van een domeintabel}
}
\value{
Een vector met kolomnamen
}
\description{
Deze functie zoekt op welke kolommen een domeintabel heeft.
}
\examples{
\dontrun{

dom_kolommen("MonsterType")

}
}
