% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dom_overzicht.R
\name{dom_overzicht}
\alias{dom_overzicht}
\title{Overzicht van alle domeintabellen}
\usage{
dom_overzicht(peildatum = Sys.Date())
}
\arguments{
\item{peildatum}{Date of een character die omgezet kan worden in een Date met
\code{lubridate::as_date()}. De peildatum filtert de output om alleen geldige
domeintabellen op de peildatum weer te geven. Gebruik \code{NULL} om alle
domeintabellen ongeacht de geldigheid weer te geven.}
}
\value{
Een tibble met een overzicht van alle domeintabellen. Het overzicht
bevat de volgende kolommen:
\itemize{
\item domeintabel - Naam van de domeintabel
\item domeintabelsoort - Het type domeintabel
\item wijzigingsdatum - Datum van de laatste wijziging van de tabel.
\item begin_geldigheid - Datum van het begin van de geldigheid van de domeintabel
\item eind_geldigheid - Datum van het eind van de geldigheid van de domeintabel
\item kolommen - Een vector met de kolomnamen van de domeintabel.
\item guid - De guid van de domeintabel
}
}
\description{
Deze functie geeft een overzicht van alle beschikbare domeintabellen,
inclusief historische tabellen, op www.aquo.nl.
}
\section{Caching}{
 Deze functie maakt gebruik van cachging voor het
optimaliseren van snelheid en om de aquo-server niet onnodig te belasten.
Standaard wordt de map \code{tempdir()} gebruikt als cache. Deze map wordt na
elke R-sessie verwijderd. Voor een persistente cache kan een zelfgekozen map
worden gebruikt. Deze map kan worden ingesteld met
\code{options(aquodom.cache_dir = "mijn_cache_dir")}. \strong{Let op} Deze cache wordt
niet automatisch gewist. Dit kan ertoe leiden dat deze functie met verouderde
data werkt. Deze optie dient dus met enige voorzichtigheid gebruikt te worden.
}

\examples{
\dontrun{

dom_overzicht()
dom_overzicht(peildatum = Sys.Date())
dom_overzicht(peildatum = "2021-04-05")

}
}
