% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dom_save.R
\name{dom_save}
\alias{dom_save}
\title{Opslaan domeintabel op schijf}
\usage{
dom_save(
  naam,
  bestandsnaam = paste(Sys.Date(), naam),
  map = NULL,
  bestandstype = c("xlsx", "csv"),
  peildatum = Sys.Date()
)
}
\arguments{
\item{naam}{Naam van een domeintabel.}

\item{bestandsnaam}{Naam van het bestand om op te slaan. Eventueel als volledig pad. Default is de datum
gevolgd door de naam van de domeintabel. De bestandsnaam mag zonder extensie worden opgegeven.}

\item{map}{Naam van de map. De map moet bestaan. Optioneel.}

\item{bestandstype}{"xlsx" of "csv"  Opslaan als xlsx of als csv-bestand. Default is xlsx.}

\item{peildatum}{De peildatum voor de domeintabel. Default is de huidige datum.}
}
\value{
Slaat de domeintabel op de schijf. De domeintabel zelf wordt onzichtbaar geretourneerd.
}
\description{
Met deze functie is het mogelijk om een domeintabel op te slaan als .xlsx-bestand of als .csv-bestand.
}
\examples{

\dontrun{
 dom_save("MonsterType")
 dom_save("MonsterType", bestandsnaam = "test.csv", map = "data", bestandstype = "csv")

}
}
\seealso{
Other domeintabellen: 
\code{\link{dom_overzicht}()},
\code{\link{dom}()}
}
\concept{domeintabellen}
