% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dom_helpers.R
\name{is_domeintabel}
\alias{is_domeintabel}
\title{Check domeintabelnamen}
\usage{
is_domeintabel(namen)
}
\arguments{
\item{namen}{Character vector met namen van domeintabellen.}
}
\value{
Logical vector
}
\description{
Checkt of een namen geldige domeintabelnamen zijn.
}
\examples{
\dontrun{

is_domeintabel("MonsterType")
is_domeintabel("Domeintabel")

}
}
